Move logic to own hook
