+++ frontend/src/App.tsx
@@ -1,36 +1,13 @@
-import { useRef, useState, useCallback } from "react";
-import { Chess } from "chess.js";
 import { Chessboard, type PieceDropHandlerArgs } from "react-chessboard";
+import { useChessGame } from "./hooks/useChessGame";
 import "./App.css";
 
 export default function Board() {
-  const gameRef = useRef(new Chess());
-  const [fen, setFen] = useState(gameRef.current.fen());
-  const [turn, setTurn] = useState<"w" | "b">(gameRef.current.turn());
+  const { fen, turn, makeMove } = useChessGame();
 
-  const onPieceDrop = useCallback(
-    ({ sourceSquare, targetSquare, piece }: PieceDropHandlerArgs): boolean => {
-      const game = gameRef.current;
-  
-      const isWhitePawn = piece?.pieceType.toLowerCase() === "wp";
-      const isBlackPawn = piece?.pieceType.toLowerCase() === "bp";
-      const isPromotion =
-        (isWhitePawn && targetSquare!.endsWith("8")) ||
-        (isBlackPawn && targetSquare!.endsWith("1"));
-
-      const move = game.move({
-        from: sourceSquare,
-        to: targetSquare!,
-        promotion: isPromotion ? "q" : undefined,
-      });
-
-      if (!move) return false;
-      setFen(game.fen());
-      setTurn(game.turn());
-      return true;
-    },
-    []
-  );
+  const onPieceDrop = ({ sourceSquare, targetSquare, piece }: PieceDropHandlerArgs) => {
+    return makeMove(sourceSquare, targetSquare!, piece?.pieceType);
+  };
 
   return (
     <div className="board">
